Add Fazz adoption.yml
`date`:
Dates from their blog posts need to be extracted from SEO crawler info

This is the public release date according to the `source`:
2020-10-05 

But the pilot started in Feb with the first contract deployment in 
2020-03-25 

See:
https://etherscan.io/tx/0xf15487117214c774de6a99b9ca57551ead74b97f257716f246f9f61ab25c51a3

Which was first announced in 2020-03-26 here:
https://www.straitsx.com/blog-post/straitsx-2020-updates

I used the date of the public release, feel free to edit as you see fit

  `entities`
StraitsX, the issuer of the stablecoin, was originally a subsidiary of fintech Xfers, which merged with fintech Payfazz in 2021 to become Fazz
https://techcrunch.com/2021/03/03/payfazz-invests-30m-in-xfers-as-the-two-southeast-asian-fintechs-form-fazz-financial-group/

I'm considering StraitsX as cryptonative, considering the portfolio of products they offer, and using the parent entity Fazz in this case 

Feel free to edit as you see fit 

  products:
This announcement is only for XSGD from 2020. Nowadays they also have XIDR, launched in 2021, and XUSD, launched in 2024. Should I add these entries separately instead?

Right now the products is shown like this:
    - XSGD Stablecoin
    - XIDR Stablecoin
    - XUSD Stablecoin

Feel free to format it in a different way if it seems weird/repetitive. 

Let me know how to handle such cases in the future when there are multiple products of the same type (funds, stablecoins, etc) on the same entry

Other possible alternatives for this entry could be:
    - XSGD, XIDR, XUSD Stablecoin 

I guess it might depend on how it renders on the website

  `chains`
  Arbitrum was not in the original release and it only started being supported this year. Should I add a separate entry for the launch on Arbitrum?
   # - Polygon